\name{nations}
\alias{nations}
\docType{data}
\title{Dataset 'nations'}
\description{Dataset containing data on 207 countries (from 'nations.txt' in Rcmdr, with missing values imputed)
}
\usage{data(nations)}
\format{
  A data frame with 207 observations on the following variables:
  \describe{
    \item{Country}{Name of the country}
    \item{TFR}{Total Fertility Rate}
    \item{contraception}{Rate of women making use of contraceptive means}
    \item{infant.mortality}{Infant mortality rate}
    \item{GDP}{Gross Domestic Product ($ per capita)}
    \item{region}{Continent (description)}
    \item{Continent}{Continent (code)}
  }
}
\source{Rcmdr package
}
\keyword{datasets}



